Rename SecondByDowning -> RevisedDowning.
I have moved this to it's own module/file.
